+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `remember_per_file_suffix` option. Default to `false`.
+
 ## 0.1.2 - 2025-09-19
 
 - Add `notify_on_switch` option. Default to `false`.